move loading view state to own view model, so we're more free how to show it, and we can better reuse it
