+++ src/domain/SessionPickerViewModel.js
@@ -1,8 +1,5 @@
 import {SortedArray} from "../observable/index.js";
 import {EventEmitter} from "../utils/EventEmitter.js";
-import {LoadStatus} from "../matrix/SessionContainer.js";
-import {SyncStatus} from "../matrix/Sync.js";
-import {loadLabel} from "./common.js";
 
 class SessionItemViewModel extends EventEmitter {
     constructor(sessionInfo, pickerVM) {
@@ -100,68 +97,6 @@ class SessionItemViewModel extends EventEmitter {
     }
 }
 
-class LoadViewModel extends EventEmitter {
-    constructor({createSessionContainer, sessionCallback, sessionId}) {
-        super();
-        this._createSessionContainer = createSessionContainer;
-        this._sessionCallback = sessionCallback;
-        this._sessionId = sessionId;
-        this._loading = false;
-    }
-
-    async _start() {
-        try {
-            this._loading = true;
-            this.emit("change", "loading");
-            this._sessionContainer = this._createSessionContainer();
-            this._sessionContainer.startWithExistingSession(this._sessionId);
-            this._waitHandle = this._sessionContainer.loadStatus.waitFor(s => {
-                this.emit("change", "loadStatus");
-                // wait for initial sync, but not catchup sync
-                const isCatchupSync = s === LoadStatus.FirstSync &&
-                    this._sessionContainer.sync.status === SyncStatus.CatchupSync;
-                return isCatchupSync ||
-                    s === LoadStatus.Error ||
-                    s === LoadStatus.Ready;
-            });
-            try {
-                await this._waitHandle.promise;
-            } catch (err) {
-                // swallow AbortError
-            }
-            if (this._sessionContainer.loadStatus.get() !== LoadStatus.Error) {
-                this._sessionCallback(this._sessionContainer);
-            }
-        } catch (err) {
-            this._error = err;
-        } finally {
-            this._loading = false;
-            this.emit("change", "loading");
-        }
-    }
-
-    get loading() {
-        return this._loading;
-    }
-
-    goBack() {
-        if (this._sessionContainer) {
-            this._sessionContainer.stop();
-            this._sessionContainer = null;
-            if (this._waitHandle) {
-                this._waitHandle.dispose();
-            }
-        }
-        this._sessionCallback();
-    }
-
-    get loadLabel() {
-        const sc = this._sessionContainer;
-        return loadLabel(
-            sc && sc.loadStatus,
-            sc && sc.loadError || this._error);
-    }
-}
 
 export class SessionPickerViewModel extends EventEmitter {
     constructor({storageFactory, sessionInfoStorage, sessionCallback, createSessionContainer}) {
@@ -193,7 +128,11 @@ export class SessionPickerViewModel extends EventEmitter {
         const sessionVM = this._sessions.array.find(s => s.id === id);
         if (sessionVM) {
             this._loadViewModel = new LoadViewModel({
-                createSessionContainer: this._createSessionContainer,
+                createAndStartSessionContainer: () => {
+                    const sessionContainer = this._createSessionContainer();
+                    sessionContainer.startWithExistingSession(sessionVM.id);
+                    return sessionContainer;
+                },
                 sessionCallback: sessionContainer => {
                     if (sessionContainer) {
                         // make parent view model move away
@@ -203,8 +142,7 @@ export class SessionPickerViewModel extends EventEmitter {
                         this._loadViewModel = null;
                         this.emit("change", "loadViewModel");
                     }
-                },
-                sessionId: sessionVM.id,
+                }
             });
             this._loadViewModel.start();
             this.emit("change", "loadViewModel");
